Get pagination working with Commission report
 - Many syntax changes for aslant: no-magic-numbers, various spaces and ;s
 - Added check that we're within our request limit before every paginated
"subrequest"
 - Rewrote the logic for determining "offset" index. Previously, we assumed
that the first page had 200 items because that's what we want. In fact, it
usually had 100 items, the MB default when we don't specify a Limit with the
initial request. So, our code was requesting items 0-99, then 200-399, then
400-599, etc. Now, this code accepts the MB-supplied length of that first
Request and correctly calculates offsets from that. So, 0-99, then 100-299,
then 300-499, etc.